backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-multipart
pydantic
pydantic-settings
python-dotenv
